Prevent potential int overflow on devices with 2048MB available heap

+++ picasso/src/main/java/com/squareup/picasso/Utils.java
@@ -293,7 +293,7 @@ final class Utils {
       memoryClass = ActivityManagerHoneycomb.getLargeMemoryClass(am);
     }
     // Target ~15% of the available heap.
-    return 1024 * 1024 * memoryClass / 7;
+    return (int) (1024L * 1024L * memoryClass / 7);
   }
 
   static boolean isAirplaneModeOn(Context context) {
